Remove HTMLUnit tests.
