+++ plugin/pom.xml
@@ -4,7 +4,7 @@
   <parent>
     <groupId>org.jvnet.hudson.plugins</groupId>
     <artifactId>analysis-pom</artifactId>
-    <version>5.17.0</version>
+    <version>5.18.0</version>
     <relativePath />
   </parent>
 
@@ -29,7 +29,7 @@
     <analysis-model-tests.version>10.9.0</analysis-model-tests.version>
 
     <forensics-api-plugin.version>1.7.0</forensics-api-plugin.version>
-    <plugin-util-api.version>2.13.0</plugin-util-api.version>
+    <plugin-util-api.version>2.14.0-rc461.1e8fc3a_a_ce9d</plugin-util-api.version>
 
     <data-tables-api.version>1.11.3-6</data-tables-api.version>
     <echarts-api.version>5.2.2-2</echarts-api.version>
