+++ plugin/src/test/java/io/jenkins/plugins/analysis/core/testutil/IntegrationTestWithJenkinsPerSuite.java
@@ -1,9 +1,7 @@
 package io.jenkins.plugins.analysis.core.testutil;
 
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.jvnet.hudson.test.JenkinsRule;
-import org.jvnet.hudson.test.JenkinsRule.WebClient;
 
 /**
  * Base class for integration tests in Jenkins. Sub classes will get a new and fresh Jenkins instance for each test
@@ -11,7 +9,7 @@ import org.jvnet.hudson.test.JenkinsRule.WebClient;
  *
  * @author Ullrich Hafner
  */
-public abstract class IntegrationTestWithJenkinsPerSuite extends IntegrationTest {
+public abstract class IntegrationTestWithJenkinsPerSuite extends WarningsIntegrationTest {
     /** Jenkins rule per suite. */
     @ClassRule
     public static final JenkinsRule JENKINS_PER_SUITE = new JenkinsRule();
@@ -19,20 +17,5 @@ public abstract class IntegrationTestWithJenkinsPerSuite extends IntegrationTest
     @Override
     protected JenkinsRule getJenkins() {
         return JENKINS_PER_SUITE;
-    }
-
-    private static WebClient noJsWebClient;
-    private static WebClient jsEnabledClient;
-
-    /** Creates two web clients, one with and one without JS support. */
-    @BeforeClass
-    public static void createWebClients() {
-        noJsWebClient = create(JENKINS_PER_SUITE, false);
-        jsEnabledClient = create(JENKINS_PER_SUITE, true);
-    }
-
-    @Override
-    protected WebClient getWebClient(final JavaScriptSupport javaScriptSupport) {
-        return javaScriptSupport == JavaScriptSupport.JS_DISABLED ? noJsWebClient : jsEnabledClient;
     }
 }
